Target Blank replacement
To accommodate the new XHTML 1.1 update and to avoid having to type target blank per href, I have alleviated the pain of external links opening on new tabs with this JavaScript update to our scripts file. Notice how this is only my contribution and not the whole code. Credit goes to "badly drawn toys".